+++ apps/backend-ops/src/firestoreMigrations/index.ts
@@ -26,7 +26,7 @@ import * as v0025 from './0025';
 import * as v0026 from './0026';
 import * as v0027 from './0027';
 import * as v0028 from './0028';
-import * as tmpV0028 from './temporary-0028';
+import * as v0029 from './0029';
 
 import { Firestore, Storage } from '../admin';
 
@@ -66,7 +66,8 @@ export const MIGRATIONS = {
   25: v0025,
   26: v0026,
   27: v0027,
-  28: v0028
+  28: v0028,
+  29: v0029,
 };
 
 export const VERSIONS_NUMBERS = Object.keys(MIGRATIONS).map(s => parseInt(s, 10));
